refactor: move database read functions to same file
